refactor : modify fields of JSON Classes

+++ src/main/kotlin/net/wavem/uvc/rms/gateway/location/domain/position/LocationPosition.kt
@@ -3,6 +3,6 @@ package net.wavem.uvc.rms.gateway.location.domain.position
 data class LocationPosition(
     val xpos : Double?,
     val ypos : Double?,
-    val heading : Int?
+    val heading : Double?
 ) {
 }
